pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rsf"
version = "0.1.0"
description = "Replacement for Step Functions — YAML DSL, code gen, and visual tooling for Lambda Durable Functions"
license = "Apache-2.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "typer>=0.9",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "sse-starlette>=1.0",
    "boto3>=1.28",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",
]

[project.scripts]
rsf = "rsf.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/rsf"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
